+++ gtars/src/tokenizers/meta_tokenizer.rs
@@ -6,5 +6,6 @@ use crate::common::models::Universe;
 
 pub struct MetaTokenizer {
     pub universe: Universe,
-    tree: HashMap<String, Lapper<u32,u32>>
+    tree: HashMap<String, Lapper<u32,u32>>,
+    secondary_trees: Option<Vec<HashMap<String, Lapper<u32, u32>>>>,
 }
