+++ gtars/src/uniwig/mod.rs
@@ -1,3 +1,16 @@
+//!
+//! # Uniwig BigWig generator
+
+//! The `uniwig` module is responsible for generating three BigWig output files based on a set of bed files. The generated files include a track for the start coordinates, a track for the end coordinates, and a track for the core coordinates.
+//!
+//! ## Under Construction
+//!
+//! This module is currently under construction. Stay tuned for more updates.
+//!
+//! ![Construction Sign](https://www.signoutfitters.com/images/products/detail/Workers_Symbol_Construction_Sign.png)
+//!
+//! 
+
 pub fn run_uniwig() {
     println!("Im running.")
 }
